% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synchr.analyze.R
\name{synchr.analyze}
\alias{synchr.analyze}
\title{Custom function to run cross-correlation analysis on dyadic time series data.}
\usage{
synchr.analyze(data, lag.max = 1, window = 0, by = 0)
}
\arguments{
\item{data, }{An optional data frame, list or environment.}

\item{lag.max, }{Number of lag periods. Must be a non-negative integer.}

\item{window, }{Window size. Must be a non-negative integer.}

\item{by, }{Overlap value for specifying overlapping windows.}
}
\description{
synchr.analyze takes data objects with columns time, series 1, and series 2 and performs cross-correlation or windowed cross-correlation analysis.
}
