% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synchr.barplots.R
\name{synchr.barplots}
\alias{synchr.barplots}
\title{Plots correlation coefficients.}
\usage{
synchr.barplots(corr_data_list)
}
\arguments{
\item{corr_data_list, }{list containing one coefficient per dyad at lag-1, lag0, and lag1.}
}
\description{
Plots correlation coefficients for each dyad across lag-1, lag0,lag1.
}
