% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synchr.classify.R
\name{synchr.classify}
\alias{synchr.classify}
\title{Categorizes cross-corrrelations to reflect synchrony style.}
\usage{
synchr.classify(corr_data_list, threshold = 0.1)
}
\arguments{
\item{corr_data_list, }{list containing one coefficient per dyad at lag-1, lag0, and lag1.}

\item{threshold, }{Threshold for what constitutes a "significant" correlation coefficient. Default is 0.10.}
}
\description{
synchr.classify uses correlation coefficients across lag-1, lag0, and lag1 to categorize dyads as either "mutual adaptors", "leader-follower", or "leader-leader."
}
