% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synchr.read.R
\name{synchr.read}
\alias{synchr.read}
\title{Custom function to read in data files into a list}
\usage{
synchr.read(data)
}
\arguments{
\item{data, }{An optional data frame, list or environment.}
}
\description{
synchr requires data to be in list format. This function allows the user to
read in all dyadic data files from one folder into a list. Data file must be in .csv format and must include a header.
}
