% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synchr.ts.R
\name{synchr.ts}
\alias{synchr.ts}
\title{Dyadic Time Series Plots.}
\usage{
synchr.ts(raw_data_list)
}
\arguments{
\item{datalist, }{list containing dyadic data.}
}
\description{
Generates simple time series plot for each element in list.
}
